Normalize Rust test file names
