+++ crates/node/src/processor/tests/test_registry.rs
