+++ crates/node/src/tests/native/mod.rs
@@ -6,7 +6,9 @@ use crate::processor::Processor;
 use crate::processor::ProcessorBuilder;
 use crate::processor::ProcessorConfig;
 #[cfg(feature = "snark")]
-pub mod snark;
+pub mod test_snark;
+
+mod test_duplicate_namespace;
 
 const TEST_DHT_FINGER_TABLE_SIZE: usize = 8;
 
@@ -30,25 +32,4 @@ pub async fn prepare_processor() -> Processor {
         .dht_finger_table_size(TEST_DHT_FINGER_TABLE_SIZE);
 
     procssor_builder.build().unwrap()
-}
-
-/// Namespace collision: registering two protocols on the same namespace is rejected, so a
-/// later extension cannot silently shadow an earlier one (one of the reviewer-requested
-/// lifecycle guards).
-#[tokio::test]
-async fn duplicate_namespace_registration_is_rejected() {
-    use std::sync::Arc;
-
-    use crate::extension::protocols::echo::Echo;
-    use crate::extension::protocols::echo::EchoShell;
-    use crate::provider::Provider;
-
-    let provider = Provider::from_processor(Arc::new(prepare_processor().await));
-    provider
-        .register_protocol(Echo, EchoShell)
-        .expect("first registration on a fresh namespace succeeds");
-    assert!(
-        provider.register_protocol(Echo, EchoShell).is_err(),
-        "a second registration on the same namespace must error, not silently overwrite"
-    );
 }
