+++ crates/node/src/tests/native/test_snark.rs
