+++ crates/node/src/tests/wasm/mod.rs
@@ -1,7 +1,7 @@
-pub mod browser;
-pub mod processor;
+pub mod test_browser;
+pub mod test_processor;
 #[cfg(feature = "snark")]
-pub mod snark;
+pub mod test_snark;
 use std::sync::Arc;
 
 use rings_core::ecc::SecretKey;
