+++ crates/node/src/tests/wasm/test_snark.rs
